Most of the class functionality set up for database content to be displayed on the website
Most of the class functionality set up for database content to be
displayed on the website. SuperCategory stores all Category objects so
that will be the one to call if all products will be shown, then
Category class holds all SubCategory objects pertaining to a specific
category, and same with SubCategory in relation to Product class. All
the classes have an array of the products themselves so that you can
display all by calling getProductArray() method when in a specific
instance of a class.

Then the website has datatables which are pulling from these new classes
to display the information requested in the xhtml code. Can't interact
with the displayed information but it is from the database so it's a
start!

I did not remove the old tables, but manageorder.xhtml
manageproducts.xhtml and manageusers.xhtml have the datatables in them.

Everything works for me so hopefully there's no issues for anyone else,
this summary is getting long as shit >.<