Can change password of first user for now
Only works for the user listed as first for some reason, and change
password page doesn't check to see if passwords are matching but maybe
someone can expand off what I did.

commandButton inside an <h:form> is what did the trick, then you can use
an action listener to get the value of the data being inserted into the
rows by calling the right method. I don't know why it only works for the
first button. But it does update the database in the correct spot so
yeah.