Products can be edited + manageusers improved
Products can now be edited and removed on the site and the changes will
be reflected in the database. Improved the user management on the site
so that password is checked to be matching when changed and delete users
is done on one page. The delete and edit buttons on both pages have 2
different states where they are hidden or not. Plus the data tables
automatically refresh when the changes are saved by clicking the button.

Only supercategory should be used for now, I need to fix up category and
subcategory still.